pyproject.toml
[project]
name = "ticktick-mcp"
version = "1.0.0"
description = "General-purpose TickTick MCP server for Claude"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
